refactor(sandbox): move csghubsdk under csghub
